+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Breaking Changes
 - Exceptions module has been moved to [PowerUtils.GuardClauses.Validations](https://github.com/TechNobre/PowerUtils.GuardClauses.Validations) project so it can be used individually;
 - ErrorCodes module has been moved to [PowerUtils.Validations.Primitives](https://github.com/TechNobre/PowerUtils.Validations.Primitives) project so it can be used individually;
+- Error codes for status code 409-Conflict changed from `ALREADY_EXISTS` to `DUPLICATED`;